🍾 Convert src/ to Typescript
