+++ src/__tests__/with-target-value.spec.ts
@@ -1,4 +1,4 @@
-import withTargetValue from '../with-target-value.js';
+import withTargetValue from '../with-target-value';
 
 
 describe('withTargetValue', () => {
