--- src/index.js
@@ -1,4 +1,0 @@
-export { default as withPreventDefault } from './with-prevent-default.js';
-export { default as withStopPropagation } from './with-stop-propagation.js';
-export { default as withStopImmediatePropagation } from './with-stop-immediate-propagation.js';
-export { default as withTargetValue } from './with-target-value.js';
