+++ src/with-prevent-default.ts
@@ -1,4 +1,4 @@
-import withEventFactory from './helpers/with-event-factory.ts';
+import withEventFactory from './helpers/with-event-factory';
 
 /**
  *  Accepts a function, to be called after calling event.preventDefault
@@ -6,6 +6,6 @@ import withEventFactory from './helpers/with-event-factory.ts';
  *  @param {Function} - function to be called after event.preventDefault
  *  @return {Function} - function which returns userland event handler
  */
-const withPreventDefault = withEventFactory('preventDefault');
+const withPreventDefault : Function = withEventFactory('preventDefault');
 
 export default withPreventDefault;
