+++ src/with-stop-immediate-propagation.ts
@@ -1,4 +1,4 @@
-import withEventFactory from './helpers/with-event-factory.ts';
+import withEventFactory from './helpers/with-event-factory';
 
 /**
  *  Accepts a function, to be called after calling event.stopImmediatePropagation
@@ -6,6 +6,6 @@ import withEventFactory from './helpers/with-event-factory.ts';
  *  @param {Function} - function to be called after event.stopImmediatePropagation
  *  @return {Function} - function which returns userland event handler
  */
-const withStopImmediatePropagation = withEventFactory('stopImmediatePropagation');
+const withStopImmediatePropagation : Function = withEventFactory('stopImmediatePropagation');
 
 export default withStopImmediatePropagation;
